A5/requirements.txt
fastapi
uvicorn
python-multipart
librosa
numpy
soundfile
scipy
pytest
pytest-cov
matplotlib
httpx
certifi
tqdm
opencv-python
ultralytics
google-generativeai>=0.8.0
jinja2
aiofiles
python-dotenv